refactor (LottoGame.java) : 패키지 이동 (root 도메인으로 사용)
